Downloads? Bare-bones
` impl Future<Output = ()> + use<'a> `

` use<'a> `

Yeah? It makes sense, but it is kinda unexpected.
Would this be more intuitive syntax?

` impl<'a> Future<Output = ()> `

Maybe, but I'm not too bothered.
What's kinda-dumb though, is that its tooltip in rust-analyzer does not mention this about the "use" keyword.

Anyways, theoretically you have the basic structure as to how these downloads are going to work.
Beyond the todo!s it would also make sense to have somekind of logging or visual progress indicator for each download; Something for the future.