await download
I think this thing technically works now.

I removed the comment inside input.rs.
The macro will always be used in a way that means it is only useful to a specific input; It is a choice between paired files or extra nesting, I would prefer extra nesting in this case; You could always make the option for either though.