async vertical
The previous structure halted a ticket's progress at each step, but allowed every ticket to execute that step  asynchronously.

So if you had one slow ticket it'd halt the tickets that had already completed their work at that step in the process.

Dumb, but also it makes it awkward to have a concurrent download limit.