Chaos and revelation
<https://users.rust-lang.org/t/an-impl-trait-contained-in-a-slice-is-typed-differently/123184/2>

The strategy thought out in the previous commit doesn't work; I thought of a few more strategies beyond that, but they all had somekind of flaw to them that made them very non-ideal.

I think I have it now though.

Start a thread per download.
When the thread finishes,
It goes back to pop a ticket off the stack, and then does its thing again.
So the stack well be held in an Arc? I think.

Under that idea, the attempted strategy wasn't even multi-threaded was it?
It was just polling all of them on a single thread in a loop?
Maybe...