reception
reception is currently "blocking" because the system will wait for the all the file reads to finish before making any requests.

A non-blocking implementation would read list in chunks and make requests as it does so; You would only care to do this for very large files, so it will go un-implemented for the foreseeable future.