as closures
Are they less performant? It seems that they need extra in this context, when compared to just having a static trait.

It is nice to be able to just define a function and have it work though; No need for an impl block.

I won't use them, it's nice to have it as dead code though.