req, hmm...
In retrospect, decoupling as I did doesn't get me anywhere.
Pointless?
or... we need to go further?
If it returned the response text and a locatch err then we'd be done.

Although, I don't think there'd be a point either way would there?
All we'd be testing at that point is the unwrapping and deserializing of some JSON.